added authentication to journal endpoints

+++ src/main/java/com/kalyani/journalApp/service/JournalEntryService.java
@@ -8,6 +8,7 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -23,17 +24,16 @@ public class JournalEntryService {
 
     @Transactional
     public void save(JournalEntry journalEntry, String userName) {
-     try{
-         User user = userService.getUserByName(userName);
-         JournalEntry saved = journalEntryRepo.save(journalEntry);
-         user.getJournalEntries().add(saved);
-       //  user.setUserName(null);
-         userService.save(user);
-     }
-     catch(Exception e){
-         System.out.println(e.getMessage());
-         throw new  RuntimeException("An error occurred while saving the journal entry",e);
-     }
+        try {
+            User user = userService.getUserByName(userName);
+            JournalEntry saved = journalEntryRepo.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            //  user.setUserName(null);
+            userService.save(user);
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            throw new RuntimeException("An error occurred while saving the journal entry", e);
+        }
     }
 
     public void saveEntry(JournalEntry journalEntry) {
@@ -49,8 +49,22 @@ public class JournalEntryService {
         return journalEntryRepo.findById(id);
     }
 
-    public void delteById(ObjectId id) {
-        journalEntryRepo.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed=false;
+        try {
+            User user = userService.getUserByName(userName);
+             removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.save(user);
+                journalEntryRepo.deleteById(id);
+            }
+
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while deleting the journal entry", e);
+        }
+        return removed;
+
     }
 
 }
